wip: splitting code in solve() into different kernels

+++ test.cu
@@ -1,19 +1,20 @@
 #include <cblas.h>
 #include <cuda_runtime.h>
+
 #include <chrono>
 #include <string>
 
 #include "include/analysis.h"
-#include "include/dynamics.h"
+#include "include/dynamics.cuh"
 #include "include/elastoplastic.h"
 #include "include/mesh.h"
 #include "include/physics.h"
 #include "include/tetrahedral.h"
 #include "include/wall.h"
 
-int main(int argc, char *argv[])
-{
-  using T = double;
+int main(int argc, char *argv[]) {
+  // using T = double;
+  using T = float;
   using Basis = TetrahedralBasis<T>;
   using Quadrature = TetrahedralQuadrature;
   using Physics = NeohookeanPhysics<T>;
@@ -27,8 +28,8 @@ int main(int argc, char *argv[])
   Mesh<T> tensile;
 
   // Material Properties
-  T E = 200E6;  // Pa
-  T rho = 7800; // kg/m3
+  T E = 200E6;   // Pa
+  T rho = 7800;  // kg/m3
   T nu = 0.25;
   T beta = 0.0;
   T H = 10;
